+++ actionview/test/template/log_subscriber_test.rb
@@ -39,7 +39,7 @@ class AVLogSubscriberTest < ActiveSupport::TestCase
 
   def set_view_cache_dependencies
     def @view.view_cache_dependencies; []; end
-    def @view.fragment_cache_key(*); "ahoy `controller` dependency"; end
+    def @view.combined_fragment_cache_key(*); "ahoy `controller` dependency"; end
   end
 
   def test_render_file_template