+++ activesupport/CHANGELOG.md
@@ -1,3 +1,9 @@
+*   Add support for versioned cache entries. This enables the cache stores to recycle cache keys, greatly saving
+    on storage in cases with frequent churn. Works together with the separation of #cache_key and #cache_version
+    in Active Record and its use in Action Pack's fragment caching.
+
+    *DHH*
+
 *   Pass gem name and deprecation horizon to deprecation notifications.
 
     *Willem van Bergen*